Improve MAC layer

+++ App/bitrev.h
@@ -4,7 +4,7 @@
 #include <inttypes.h>
 #include "macros.h"
 
-_inline_ uint16_t __REV16(uint16_t brev) {
+_inline_ uint32_t __REV16(uint32_t brev) {
   return ((brev & 0xff00) >> 8) | ((brev & 0xff) << 8);
 }
